requirements.txt
numpy==1.23.3
pandas==1.4.4
python_igraph==0.11.2
scikit_learn==1.1.3
scipy==1.9.3
wget==3.2
rarfile==4.0
